Integrate CircleCI (#48)
* Add circleci

* Removed minimal templates, line_profiler has no concept of minimal builds

* manylinux build is now on circleci, still wip

* remove 2.7 from travis

* Add kernprof to test coverage

* Use new GPG Key. Testing if dist is preserved between steps in a workflow

* disable travis

* combine sign and deploy stage, hack in branch conditionality

* switch readme to circleci

+++ clean.sh
@@ -14,7 +14,10 @@ rm -rf wheelhouse
 rm -rf pip-wheel-metadata
 rm -rf htmlcov
 
-rm distutils.errors || echo "skip rm"
+
+if [ -f "distutils.errors" ]; then
+    rm distutils.errors || echo "skip rm"
+fi
 
 CLEAN_PYTHON='find . -regex ".*\(__pycache__\|\.py[co]\)" -delete || find . -iname *.pyc -delete || find . -iname *.pyo -delete'
 bash -c "$CLEAN_PYTHON"
